% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_value_high.R
\name{card_value_high}
\alias{card_value_high}
\title{Return the value of a single card under "high" assumption, i.e. A = 14}
\usage{
card_value_high(card)
}
\arguments{
\item{card}{The number (without suit) of a single card}
}
\value{
a number from 2,3,....,14
}
\description{
Return the value of a single card under "high" assumption, i.e. A = 14
}
\examples{
card_value_high("A")
}
