% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_value_low.R
\name{card_value_low}
\alias{card_value_low}
\title{Return the value of a single card under "low" assumption, i.e. A = 1}
\usage{
card_value_low(card)
}
\arguments{
\item{card}{The number (without suit) of a single card}
}
\value{
a number from 1,2,....,13
}
\description{
Return the value of a single card under "low" assumption, i.e. A = 1
}
\examples{
card_value_low("A")
}
