% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_validation.R
\name{hand_validation}
\alias{hand_validation}
\title{Whether a 5-card hand is valid}
\usage{
hand_validation(hand)
}
\arguments{
\item{hand}{A hand that consists of 5 cards}
}
\value{
True or False
}
\description{
Whether a 5-card hand is valid
}
\examples{
hand_validation(c("2h","3c","3s","4d","Kh"))
}
