% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_flush.R
\name{is_flush}
\alias{is_flush}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a flush and its strength if T}
\usage{
is_flush(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a flush and its strength if T
}
\examples{
is_flush(c("Ks","As","2s","3s","4s"))
}
