% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fullhouse.R
\name{is_fullhouse}
\alias{is_fullhouse}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a full house and its strength if T}
\usage{
is_fullhouse(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a full house and its strength if T
}
\examples{
is_fullhouse(c("5s","6h","5c","5h","6d"))
}
