% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_high_card.R
\name{is_high_card}
\alias{is_high_card}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a high card and its strength if T}
\usage{
is_high_card(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a high card and its strength if T
}
\examples{
is_high_card(c("5s","6h","7c","8h","Td"))
}
