% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_quads.R
\name{is_quads}
\alias{is_quads}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a quads and its strength if T}
\usage{
is_quads(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a quads and its strength if T
}
\examples{
is_quads(c("5s","5d","5h","5c","As"))
}
