% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_straight.R
\name{is_straight}
\alias{is_straight}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a straight and its strength if T}
\usage{
is_straight(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a straight and its strength if T
}
\examples{
is_straight(c("As","2s","3s","4s","5s"))
}
