% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_trips.R
\name{is_trips}
\alias{is_trips}
\title{Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a trips and its strength if T}
\usage{
is_trips(hand)
}
\arguments{
\item{hand}{A 5-card hand}
}
\value{
list(T/F, strength)
}
\description{
Return the list: list(T/F, strength) of a 5-card hand stating whether (T/F) a 5-card hand is a trips and its strength if T
}
\examples{
is_trips(c("5s","6h","5c","5h","7d"))
}
